ai_engine/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
torch>=2.2
torch-geometric>=2.5
transformers>=4.38
numpy>=1.26
scipy>=1.12
scikit-learn>=1.4
pandas>=2.2
networkx>=3.2
httpx>=0.27
python-dotenv>=1.0
